backend/requirements.txt
Django==4.2
psycopg2-binary==2.9.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1 